refactor: app.py 파일 분리
