+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,28 +1,33 @@
 ---
 name: 🐞 버그 리포트
 about: 예기치 않은 문제나 오류를 보고합니다.
-title: "[BUG] "
+title: '[BUG] '
 labels: bug
 assignees: ''
-
 ---
 
 ## 🐛 버그 설명
+
 무슨 문제가 발생했는지 설명해주세요.
 
 ## ✅ 재현 방법
+
 어떤 상황에서 문제가 발생했는지 단계별로 설명해주세요.
 
 ## 🖥️ 예상 결과
+
 정상적으로 동작했다면 어떤 결과가 나와야 하나요?
 
 ## ⚠️ 실제 결과
+
 실제로는 어떤 결과가 나왔나요?
 
 ## 🌐 환경 정보
+
 - 운영체제:
 - 브라우저 또는 실행 환경:
 - 관련 버전 (예: Python, Node.js, Java 등):
 
 ## 📸 참고 자료
+
 스크린샷이나 로그를 첨부할 수 있다면 추가해주세요.
