+++ .github/ISSUE_TEMPLATE/discussion.md
@@ -1,20 +1,23 @@
 ---
 name: 💬 논의
 about: 기능 방향성이나 아키텍처에 대해 논의합니다.
-title: "[DISCUSSION] "
+title: '[DISCUSSION] '
 labels: discussion
 assignees: ''
-
 ---
 
 ## 💬 논의 주제
+
 어떤 주제에 대해 논의하고 싶나요?
 
 ## 💡 배경 및 맥락
+
 왜 이 주제가 중요한가요? 어떤 문제가 발생하고 있나요?
 
 ## 🔀 선택지 또는 아이디어
+
 가능한 해결책이나 방향성을 나열해주세요.
 
 ## 🙋‍♀️ 논의하고 싶은 부분
+
 구체적으로 어떤 점에 대한 의견이 필요하신가요?
