+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,20 +1,23 @@
 ---
 name: ✨ 기능 구현
 about: 새로운 기능 구현을 요청하거나 제안합니다.
-title: "[FEATURE] "
+title: '[FEATURE] '
 labels: enhancement
 assignees: ''
-
 ---
 
 ## ✨ 기능 설명
+
 어떤 기능을 제안하시나요?
 
 ## 🤔 동기 또는 문제
+
 왜 이 기능이 필요한가요? 어떤 문제를 해결하나요?
 
 ## 🔧 예상 구현 방식
+
 기능이 어떻게 구현될 수 있을지 간단히 설명해주세요.
 
 ## 📝 참고 자료
+
 비슷한 기능, 관련된 레퍼런스, 문서 등이 있다면 첨부해주세요.
