+++ .github/ISSUE_TEMPLATE/question.md
@@ -1,17 +1,19 @@
 ---
 name: ❓ 질문
 about: 프로젝트 또는 코드에 대한 질문을 남깁니다.
-title: "[QUESTION] "
+title: '[QUESTION] '
 labels: question
 assignees: ''
-
 ---
 
 ## ❓ 질문 내용
+
 어떤 점이 이해되지 않거나 헷갈리시나요?
 
 ## 📄 관련 문서/코드
+
 관련된 파일이나 기능이 있다면 명시해주세요.
 
 ## 💬 기대 답변
+
 어떤 정보를 얻고 싶은가요?
