+++ doc/infra/01_aws_instance_create.md
@@ -1,10 +1,8 @@
 01_aws_instance_create.md
 
 > ❗ **이 문서는 DevOps 담당자 또는 인프라 관리자 참고용입니다.**
-> 
-> 
+>
 > **팀원은 직접 수행할 필요 없습니다.**
-> 
 
 # ⚙️AWS EC2 Instance 초기설정
 
@@ -47,4 +45,5 @@
 ---
 
 ## 작성자: 김세찬 (DevOps 담당)
+
 작성일: 2025-05-06
