+++ doc/infra/04_domain_config.md
@@ -1,9 +1,8 @@
 04_domain_config.md
+
 > ❗ **이 문서는 DevOps 담당자 또는 인프라 관리자 참고용입니다.**
-> 
-> 
+>
 > **팀원은 직접 수행할 필요 없습니다.**
-> 
 
 ## 1. 도메인 구매 홈페이지 : Gabia
 
@@ -17,22 +16,24 @@
 
 - 루트도메인
 
-| 항목 | 입력 값 |
-| --- | --- |
-| 호스트 이름 | `@`  |
-| 타입 | `A` |
-| TTL | 3600 |
-| 값(Value) | EC2 퍼블릭 IP(3.38.97.96) |
+| 항목        | 입력 값                   |
+| ----------- | ------------------------- |
+| 호스트 이름 | `@`                       |
+| 타입        | `A`                       |
+| TTL         | 3600                      |
+| 값(Value)   | EC2 퍼블릭 IP(3.38.97.96) |
+
 - 서브도메인
 
-| 항목 | 입력 값 |
-| --- | --- |
-| 호스트 이름 | `www` |
-| 타입 | `A` |
-| TTL | 3600 |
-| 값 | EC2 퍼블릭 IP(3.38.97.96) |
+| 항목        | 입력 값                   |
+| ----------- | ------------------------- |
+| 호스트 이름 | `www`                     |
+| 타입        | `A`                       |
+| TTL         | 3600                      |
+| 값          | EC2 퍼블릭 IP(3.38.97.96) |
 
 ---
 
 ## 작성자: 김세찬 (DevOps 담당)
+
 작성일: 2025-05-06
