+++ doc/reference/03_Dockere_Instruction.md
@@ -1,15 +1,14 @@
 ## 1. 팀원 도커 사용 기본 방법
 
 - 도커 컨테이너 실행
-    - docker-compose up --build
+  - docker-compose up --build
 - 도커 컨테이너 종료
-    - docker-compose down
+  - docker-compose down
 - 도커 컨테이너 접속
-    - docker exec -it <컨테이너이름> bash
+  - docker exec -it <컨테이너이름> bash
 - 도커 컨테이너 재시작
-    - docker-compose restart
-
-
+  - docker-compose restart
 
 ## 작성자: 김세찬 (DevOps 담당)
+
 작성일: 2025-05-10
