+++ doc/security/Infra_log.md
@@ -8,10 +8,12 @@ Infra_log.md
 - Public IPv4 : 3.38.97.96
 - 보안 그룹: launch-wizard-2 (sg-0a376)
 - 변경 내용:
-    - TCP 80번 (HTTP) 포트 개방 → 0.0.0.0/0
-    - 비고: 도메인([goathub.shop](http://goathub.shop/)) 웹 접근 허용 목적
 
-    ---
+  - TCP 80번 (HTTP) 포트 개방 → 0.0.0.0/0
+  - 비고: 도메인([goathub.shop](http://goathub.shop/)) 웹 접근 허용 목적
+
+  ***
 
 ## 작성자: 김세찬 (DevOps 담당)
+
 작성일: 2025-05-06
