+++ src/config.py
@@ -1,0 +1,32 @@
+import os
+
+import streamlit as st
+from dotenv import load_dotenv
+
+
+def load_dotenv_and_session(localS):
+    load_dotenv()
+    if "notion_token" not in st.session_state:
+        st.session_state.notion_token = localS.getItem("notion_token") or ""
+    if "notion_db_id" not in st.session_state:
+        st.session_state.notion_db_id = localS.getItem("notion_db_id") or ""
+    if "proxy_username" not in st.session_state:
+        st.session_state["proxy_username"] = os.getenv("WEBSHARE_PROXY_USERNAME")
+        st.session_state["proxy_password"] = os.getenv("WEBSHARE_PROXY_PASSWORD")
+
+
+def init_session():
+    default_values = {
+        "video_id": "",
+        "transcript_text": "",
+        "transcript_data": None,
+        "summary": "",
+        "summarize_clicked": False,
+        "summarizing": False,
+        "summarized": False,
+        "auto_save_to_notion": True,
+        "notion_saved": False,
+    }
+    for k, v in default_values.items():
+        if k not in st.session_state:
+            st.session_state[k] = v
